fix: remove unused job policy
Coses eveseat/seat#743
