+++ src/resources/views/character/sheet.blade.php
@@ -221,9 +221,7 @@
           @if($character->refresh_token)
           <div class="card-tools">
             <div class="input-group input-group-sm">
-              <button type="button" class="btn btn-sm btn-light" data-widget="esi-update" data-character="{{ $character->character_id }}" data-job="character.corphistory">
-                <i class="fas fa-sync" data-toggle="tooltip" title="{{ trans('web::seat.update_corp_history') }}"></i>
-              </button>
+              @include('web::components.jobs.buttons.update', ['type' => 'character', 'entity' => $character->character_id, 'job' => 'character.corphistory', 'label' => trans('web::seat.update_corp_history')])
             </div>
           </div>
           @endif
