+++ src/resources/views/layouts/app-mini.blade.php
@@ -18,7 +18,8 @@
     <!-- Theme style -->
     <link rel="stylesheet" href="{{ asset('web/css/adminlte.min.css') }}">
     <!-- Google Font -->
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700">
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css2?family=Source+Sans+Pro:ital,wght@0,300;0,400;0,700;1,400&display=swap" rel="stylesheet">
 
     <!-- Custom layout CSS -->
     @if(file_exists(public_path('custom-layout-mini.css')))
