+++ src/resources/views/layouts/app.blade.php
@@ -23,7 +23,8 @@
     <!-- Theme style -->
     <link rel="stylesheet" href="{{ asset('web/css/adminlte.min.css') }}">
     <!-- Google Font -->
-    <link href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700" rel="stylesheet">
+    <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css2?family=Source+Sans+Pro:ital,wght@0,300;0,400;0,700;1,400&display=swap" rel="stylesheet">
     <!-- SeAT CSS -->
     <link rel="stylesheet" href="{{ asset('web/css/seat.css') }}">
     <!-- SeAT Skins CSS -->
